limited features to auth users

+++ blog/resources/views/posts/show.blade.php
@@ -34,18 +34,23 @@
           </dd>
         </dl>
         <hr>
-        <div class='row'>
-          <div class='col-sm-6'>
-            {!! Html::linkRoute('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-danger btn-block')) !!}
-          </div>
-          <div class='col-sm-6'>
-            {!! Form::open(['route' => ['posts.destroy', $post->id], 'method' => 'DELETE']) !!}
 
-              {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-block']) !!}
+        @if (Auth::check())
 
-            {!! Form::close() !!}
+          <div class='row'>
+            <div class='col-sm-6'>
+              {!! Html::linkRoute('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-danger btn-block')) !!}
+            </div>
+            <div class='col-sm-6'>
+              {!! Form::open(['route' => ['posts.destroy', $post->id], 'method' => 'DELETE']) !!}
+
+                {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-block']) !!}
+
+              {!! Form::close() !!}
+            </div>
           </div>
-        </div>
+          
+          @endif
 
       </div>
